orderingsys/requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.0.3
django-crispy-forms==1.9.2
django-filter==2.2.0
django-formtools==2.2
django-import-export==2.3.0
django-redis==4.10.0
django-reversion==3.0.7
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
DjangoUeditor
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient 
odfpy==1.4.1
openpyxl==3.0.4
Pillow==6.2.1
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
six==1.15.0
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
xadmin
xlrd==1.2.0
xlwt==1.3.0